+++ src/utils/calculation_manager.c
@@ -6,20 +6,20 @@
 /*   By: nifromon <nifromon@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/25 03:20:04 by nifromon          #+#    #+#             */
-/*   Updated: 2025/04/28 23:11:02 by nifromon         ###   ########.fr       */
+/*   Updated: 2025/04/29 03:20:48 by nifromon         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../includes/cub.h"
 
 // Function to convert degree to radian
-double	cub_degtorad(double degree)
+float	cub_degtorad(float degree)
 {
 	return (degree * (M_PI / 180.0));
 }
 
 // Function to fix the angle
-double	cub_fixang(double angle)
+float	cub_fixang(float angle)
 {
 	if (angle > 359)
 		angle -= 360;
@@ -29,7 +29,7 @@ double	cub_fixang(double angle)
 }
 
 // Function to calculate the distance between two points
-double	cub_calc_dist(t_pos a, t_pos b, double angle)
+float	cub_calc_dist(t_pos a, t_pos b, float angle)
 {
 	return (cos(cub_degtorad(angle)) * (b.x - a.x)
 		- sin(cub_degtorad(angle)) * (b.y - a.y));
@@ -38,7 +38,7 @@ double	cub_calc_dist(t_pos a, t_pos b, double angle)
 
 
 // Function to convert gl colors to rgb and rgb into colors
-int	cub_convert_glrgb(double red, double green, double blue, int mode)
+int	cub_convert_glrgb(float red, float green, float blue, int mode)
 {
 	int	r;
 	int	g;
